rename package from operations to utils
