+++ deeprankcore/features/components.py
@@ -1,7 +1,7 @@
 from typing import Optional
 import numpy as np
-from deeprankcore.operations.graph import Graph
-from deeprankcore.operations.parsers import atomic_forcefield
+from deeprankcore.utils.graph import Graph
+from deeprankcore.utils.parsers import atomic_forcefield
 from deeprankcore.molstruct.variant import SingleResidueVariant
 from deeprankcore.molstruct.atom import Atom
 from deeprankcore.molstruct.residue import Residue
