+++ deeprankcore/features/contact.py
@@ -4,10 +4,10 @@ import warnings
 import numpy as np
 from scipy.spatial import distance_matrix
 from deeprankcore.molstruct.atom import Atom
-from deeprankcore.operations.graph import Graph
+from deeprankcore.utils.graph import Graph
 from deeprankcore.molstruct.pair import ResidueContact, AtomicContact
 from deeprankcore.domain import edgefeatures as Efeat
-from deeprankcore.operations.parsers import atomic_forcefield
+from deeprankcore.utils.parsers import atomic_forcefield
 
 _log = logging.getLogger(__name__)
