+++ deeprankcore/molstruct/residue.py
@@ -2,7 +2,7 @@ from typing import Optional
 import numpy as np
 from deeprankcore.molstruct.structure import Chain
 from deeprankcore.molstruct.aminoacid import AminoAcid
-from deeprankcore.operations.pssmdata import PssmRow
+from deeprankcore.utils.pssmdata import PssmRow
 
 
 class Residue:
