+++ deeprankcore/neuralnets/foutnet.py
@@ -5,7 +5,7 @@ import torch.nn.functional as F
 from torch_scatter import scatter_mean
 from torch_geometric.nn.inits import uniform
 from torch_geometric.nn import max_pool_x
-from deeprankcore.operations.community_pooling import get_preloaded_cluster, community_pooling
+from deeprankcore.utils.community_pooling import get_preloaded_cluster, community_pooling
 
 
 class FoutLayer(torch.nn.Module):
