+++ deeprankcore/utils/grid.py
