+++ deeprankcore/utils/parsers/residue.py
