+++ tests/features/test_aminoacid.py
@@ -2,8 +2,8 @@ from pdb2sql import pdb2sql
 from deeprankcore.domain.aminoacidlist import serine
 from deeprankcore.molstruct.variant import SingleResidueVariant
 from deeprankcore.features.surfacearea import add_features
-from deeprankcore.operations.graph import build_residue_graph
-from deeprankcore.operations.buildgraph import get_structure, get_surrounding_residues
+from deeprankcore.utils.graph import build_residue_graph
+from deeprankcore.utils.buildgraph import get_structure, get_surrounding_residues
 from deeprankcore.domain import nodefeatures
 from deeprankcore.features.components import add_features # noqa
