+++ tests/features/test_contact.py
@@ -4,8 +4,8 @@ import numpy as np
 from deeprankcore.molstruct.structure import Chain
 from deeprankcore.molstruct.atom import Atom
 from deeprankcore.molstruct.pair import AtomicContact, ResidueContact
-from deeprankcore.operations.graph import Edge, Graph
-from deeprankcore.operations.buildgraph import get_structure
+from deeprankcore.utils.graph import Edge, Graph
+from deeprankcore.utils.buildgraph import get_structure
 from deeprankcore.features.contact import add_features
 from deeprankcore.domain.aminoacidlist import alanine
 from deeprankcore.molstruct.variant import SingleResidueVariant
