+++ tests/features/test_surfacearea.py
@@ -5,8 +5,8 @@ from deeprankcore.molstruct.structure import PDBStructure, Chain
 from deeprankcore.molstruct.residue import Residue
 from deeprankcore.molstruct.variant import SingleResidueVariant
 from deeprankcore.features.surfacearea import add_features
-from deeprankcore.operations.graph import build_residue_graph, build_atomic_graph
-from deeprankcore.operations.buildgraph import (
+from deeprankcore.utils.graph import build_residue_graph, build_atomic_graph
+from deeprankcore.utils.buildgraph import (
     get_structure,
     get_residue_contact_pairs,
     get_surrounding_residues)
