+++ tests/models/test_graph.py
@@ -4,10 +4,10 @@ import os
 import h5py
 from pdb2sql import pdb2sql
 import numpy as np
-from deeprankcore.operations.grid import GridSettings, MapMethod
-from deeprankcore.operations.graph import Graph, Edge, Node
+from deeprankcore.utils.grid import GridSettings, MapMethod
+from deeprankcore.utils.graph import Graph, Edge, Node
 from deeprankcore.molstruct.pair import ResidueContact
-from deeprankcore.operations.buildgraph import get_structure
+from deeprankcore.utils.buildgraph import get_structure
 from deeprankcore.domain import (nodefeatures as Nfeat,
                                 edgefeatures as Efeat)
